+++ test/utils/_check-result.js
